requirements.txt
django==5.0
djangorestframework==3.14.0
drf-yasg==1.21.7
python-decouple==3.8
pillow==10.1.0
